Add dbapi functions
1. add get_nodes_by_chassis
2. add chassis_list
3. add tests for new dbapi function
4. remove object created in setUp function in test

We need chassis_list when we get /chassis,
and we need get_nodes_by_chassis when we get /chassis/chassis_id/nodes.

I remove object created in setUp function because the object created in
setUp function will always exist in the db util test is over, so the
result of test_chassis_list will be effected by the object created in
setUp.

Change-Id: Ic2a743b764761935f847de0e0cdb0a2ebcec2486