Add Pagination of collections across the API
Pagination is designed to return a subset of the larger collection while
providing a link that can be used to retrieve the next. The next link
will preserve any parameters sent in the initial request.

The patch also creates a new 'api' option group and move the existing
api configuration from the [default] section to the new [api] section.

Implements blueprint add-api-pagination

Change-Id: I3e72599344b7716a3ac88e504d018cbe7b0764b3