Add API uuid <-> id mapping
On the API side, when creating a new node or port, translate 'chassis_id'
and 'node_id' from uuid to id, internally id will be always used to do
database operations. The opposite will happen when returning data back
to the user the internal id will be translated to an uuid.

Change-Id: I84782268c4317b4eec53fdcdfc6cb2b901b22780