Upgrade tox to 1.6
tox 1.6 was released, which means that we can now take advantage of the
feature we added to it - which is using setup.py develop to install the
code into the virtualenv. The logic was taken from run_tests.sh - so the
performance issues around using tox vs. using install_venv should now be
gone.

Change-Id: Ib011a9add595696f5db5fe596db9d3d60f46d038
Fixes-Bug: #1233213