Fix policies
Policy file contained malformed content so the policy
engine failed to parse it. That was the reason of rejecting
all requests, if authentication was enabled.

This patch also updates policies to get rid of unused policies
and use GenericCheck to check for admin API.

After changes mentioned above some unused code appeared in
ironic.common.policy and so it was cleaned up.

Closes-bug: #1236371
Change-Id: Ie1dbda11561a9e7068d240a19f9fb98eae121c94