Supports paginate query for two get nodes DB APIs
In DB APIs that return a list of nodes, limiting and sorting the
output is necessary for a large number of output nodes. This is
especially important for the GUI that uses the APIs where the view
is limited. This patch adds limit, marker, sort_key and sort_dir
to get_associated_nodes and get_unassociated_nodes DB APIs.

Change-Id: Ia2c616fff65a03bace1244fe646bdf0cf5f30a29