Do not allow API to expose error stacktrace
Due to specific implementation of RPC error deserialization we have
remote exception traceback built into the error message, therefore
web server puts it to the response and sends to the client. This
patch adds Pecan hook that cuts off traceback from the error message
before it sent to the client.

Closes-Bug: #1244747
Change-Id: Ic12f151123f4ccce36d64608873eeb99eccc525a