API ports update for WSME 0.5b6 compliance
This makes use of the mandatory option of WSME, to remove some of the
custom validation code. The patch also includes a new attribute on
Ports API object called node_uuid to store the UUID of the node that
Port belongs to, once this field is set it magically converts the UUID
to the numeric ID of the node and sets the node_id attribute to be
used internally.

Change-Id: I973987de48fa008f5752861a2acfaf819f80e2ad
Partial-Bug: #1252213