Add heartbeat_interval to new 'conductor' cfg group
Add new 'conductor' config option group, to separate the various options
specific to this service which will be added in subsequent patches.

Add heartbeat_interval option to control how often a conductor updates
its status record in the database.

Fix a minor issue with the unit test associated with this method
and rename it to be more consistent with other unit test names.

Change-Id: I0ea71231ba7a6f20b94f4bca7de2a2f49bc4a428