Add 'next' link when GET maximum number of items
When GET'ing items from a collection with no limit parameter specified
in the request, the API gets the limit value from the configuration file
(default is max_limit=1000). However, in the case where there are more
than max_limit items, the API was not returning a 'next' link to retrieve
the next subset of items.

It turns out that the max_limit value wasn't being passed to the
*Collection.convert_with_links method. This fixes that issue by moving
the collection call into the same method that determines the limit value.

Change-Id: I81e734a951e4025f64a5d572656d627e4126b353
Closes-Bug: #1250804