Make chassis_uuid field of nodes optional
Currently chassis is not being used by Ironic, it's nice to have it in
our API as a pre-design for grouping nodes that shares some capabilities
(e.g nodes that shares the firmware) or just as a way of grouping the
nodes for organization purpose. It shouldn't be mandatory to specify a
chassis uuid when creating one node since it might be difficult for other
things, eg. migration from Nova. Nova doesn't have a similar abstraction,
so it's impossible to determine which chassis the node belongs to when
migrating from it. So, this patch is making the chassis_uuid field of nodes
optional.

Change-Id: I9fe657ab11d809271eea44d734a5986655a5b303