API to validate UUID parameters
The patch uses the UuidType in the wsexpose decorator to make WSME
validate whether UUIDs parameters are UUID-like strings. By doing that,
it solves the problem of allowing ports to be accessed by its address,
which is an inconsistence in our API.

As our UuidType now differs from the UuidType of WSME the patch also
removes the note to remove our custom type on the next WSME release.

Change-Id: I023e8f31d94555577c4573b4a5cf0d649d33b4fa
Closes-Bug: #1246816