PXE and SSH validate() method to check for a port
The PXE and SSH server depends on a MACAddress to work, before the
validate() method of the PXE and SSH drivers didn't check whether the
node had a port associated with it or not making it possible to start
a provision/power state change without and address that would fail. So,
this patch is improving the validate() method of the PXE and SSH driver
to check for a port.

Closes-Bug: #1272045
Change-Id: I1e8deb91558478973b280ed25bf3b90da4e2b045