Add common.service config options to sample
openstack/common/config/generator.py looks for a variable that is a list
of config options. The code in common.service was written with the list
passed as an argument to cfg.CONF.register_opts(), instead of being
assigned to a variable first, so the generator didn't detect it.

In common.service, the list of config options is now assigned to a variable.
The ironic.conf.sample file generated by tools/config/generate_sample.sh
now includes the two config options: periodic_interval and host.

Change-Id: I2f42e7f407ffc157d225652045c1bf53f65405b3
Closes-Bug: #1280064