NodeLocked returns 503 error status
The NodeLocked exception status code was incorrectly set to 409
CONFLICT.  That status is (correctly) used by client-side errors, such
as MACAlreadyExists and InvalidState, which belong in the 4xx class of
HTTP error codes.

Instead, NodeLocked is a temporary server-side error and as such should
use the 503 error code, indicating that the client should re-issue the
request unchanged after waiting for the resource state to change on the
server side.

Change-Id: I00c6bac32ac7db068f49633a576422f9ca93cfef