Fix seamicro validate() method definition
The base class definition of vendor_passthru.validate() was changed
in 48a4223f so that it also required a 'task' parameter, but
the seamicro.vendor_passthru driver had not landed yet. Between that
change and the time the seamicro driver was landed, another change
caused us to stop running unit tests on third-party drivers.

So, we didn't notice that the seamicro.vendor_passthru.validate() method
definition was incompatibly broken.

Closes-bug: #1298648

Change-Id: I9eab4c1d1be0546a0cdbf40759e88bb22f5015d2