JsonPatch add operation on existing property
Before a validation code in Ironic was preventing any 'add' operation
to the root of the document, but this create an inconsistency with
multi-level attributes (e.g extra) where 'add' operations works
interchangeable if used on existing properties, this behavior is expect
as per RFC 6902. So this patch makes it possible to use 'add' operations
to replace the values of the properties at the root of the document.

It's important to note that as per Ironic design we do not allow to add
new properties to the root of the document.

Change-Id: Ie448ec46273d7a849c135940084d72d66b9b8a53
Closes-Bug: #1278441