Document ClusteredComputeManager
Document the option needed to enable the ClusteredComputeManager in
nova.conf.

The commit a719472d added a custom compute manager to Ironic which
prevents some startups problem (e.g do not cause n-cpu to die at startup
if Ironic is not running yet) and workaround some frictions that we have
with Nova which assumes that each n-cpu is the only one responsible for
an instance (which is not true in the Ironic model).

Change-Id: I26bf5a1d3b634788044991e4524f64acf23bed9a