Wait for Neutron port updates when using SSHPower
Adds a sleep after updating Neutron from the PXEDeploy driver. This only
triggers if node is configured to use the SSHPower driver. In this case, it is
presumed to be powering on virtual machines which may present a race between
it booting and Neutron processing DHCP config changes.  This is a temporary
fix until we have better synchronization with external services.  The 15s sleep
is hard-coded currently to avoid adding a temporary config setting.

Closes-bug: #1334447.
Partial-Bug: #1300589.

Change-Id: Iee40d45d34af0c44bc14c202424e919713efb78d