Pass kwargs to ClientWrapper's call() method
As the ClientWrapper class had no unittests some were added to test
the specific part of the code for this change. The FakeClient() and
associated classes from test_driver.py were moved to a common file to
make it easier to be imported and used by the new unittests.

Partial-bug: #1316549
Related-bug: #1309719
Change-Id: I11be25a6ce981e63027073d12fbf2922896f4693