Pin RPC client version min == max
At the moment, ironic's rpcapi module is specifying a target (minimum)
RPC version of 1.0 -- however, none of the methods added after 1.0
indicate in which minor version they were added. Thus, this module
believes all methods are available in any version of the server greater
than 1.0, which is clearly untrue.

This change pins the target version to the current version,
effectively preventing rolling upgrades, which are not supported at the
moment anyway.

Change-Id: I031ccd13d1d140821daf8a0adf5913627064451c
Partial-Bug: #1317300