Add some real-world testing on DiskPartitioner
This patch adds some integration tests for DiskPartitioner by using
it on a temporary file. As a side effect, this patch adds
disk_partitioner.list_partitions() helper, which is used by the test,
but can by used for anything else later.

The new tests are currently skipped on gate-ironic-python26, as
there is no `parted` utility there.
Also these tests do not cover destroy_disk_metadata, as it can't
be executed on a file.

Change-Id: I53bf2d207b7d15a0673be6a6786c47500b9eb499
Closes-Bug: #1306153