Use metadata.create_all() to initialise DB schema
There is no need to run database migration each time, when we want to
get a new actual database schema. Running all migration scripts takes a
long time on production databases and can be not suitable for SQLite
because of Alembic limitation.

Added create_schema() method, which supposes to be used for initial
installation instead of upgrade('head').

Partial-Bug: #1347604

Change-Id: I039d955f051eda015bef2299122fd1f0fb5c46da