Nova review updates for _node_resource
In review https://review.openstack.org/#/c/111425/14 it has been
raised that if the cpu_arch is not set, we should be using Nova
helper functions instead of 'NotFound'.  Also we should be
canonicalizing the user provided config value.

Once this is approved and merged, it will be proposed back into
the Nova tree.

Change-Id: Idfe0730daa22c6550a1a15b45f481635d2eb5195