Set PYTHONHASHSEED for venv tox environment
We're seeing frequent failures of the docs Jenkins job that appear
to be caused by the tox 1.7.2 upgrade and the migration to using
PYTHONHASHSEED=random.  This overrides the environment in the venv
testenv to set PYTHONHASHSEED=0 and allows docs to build
successfully.  Note that we want to run our tests with a random seed
if possible, so this avoids disabling it for the base testenv and
only uses non-random for the testenv used by Jenkins for doc builds.

Change-Id: I665703c9a002500bc90b90cd268bd3ccff4e3534
Closes-bug: 1361295