Don't write python bytecode while testing
This patch was based on the Nova patch:
https://review.openstack.org/#/c/121044

This should reduce the number of times you need to completely delete
your tox dir because you have stale pyc files around.

In Ironic we have hit this problem a number of times, migrating to
oslo.message was a big one that caused a lot of confusion.

Closes-Bug: #1368661
Change-Id: I7dda3ad879beff1e494a36144c941a847177c2ef