Stop running check_uptodate in the pep8 testenv
As per this discussion on the mailing list:
  http://lists.openstack.org/pipermail/openstack-dev/2014-September/047160.html

We are going to stop enforcing the correctness of the sample config
file, and move to a separate non-voting check test that notifies us when
it is no longer accurate, thus prompting a human to correct it.

This addresses the issue where config options in dependent libraries
(eg. keystone) repeatedly break our gate, without removing the sample
config file that deployers benefit from.

Change-Id: Ief88a693e4967b32f5f2f71b9a683599527a5129