Iterate over glance API servers
The configuration option CONF.glance.glance_api_servers is meant to
be a list of Glance API servers, and should behave similar to its use
in Nova and Cinder: that each time a Glance API server is wanted, one
of the servers in the list is returned.

This changes CONF.glance.glance_api_servers to be a list instead of
a string. The list is randomly shuffled, and an iterator is
used to cycle through the list of servers.

This is backwards compatible and handles the case where the
user may have set [glance]/glance_api_servers = 'host:port' (or
'host1:port1, host2:port2'). The configuration library will
convert that to the corresponding list.

Closes-Bug: #1381135

Change-Id: Ic7cab502f38407aa0c282203804023dfa45b6623