Put nodes-related API in same section
The documentation for our web API (v1) has the nodes-related API
in three different sections: Nodes, NodeStates, and BootDevice.
This makes it difficult to find all the available nodes-related
API.

This puts all the nodes-related API in the same Nodes section,
removing the NodeStates and BootDevice sections.

The ConsoleInfo class/type is included in the documentation as well.

Change-Id: I3bdcb2df13861d3a098b1880062b8115855c28b6
Closes-Bug: #1386291