Fix image_info passed to IPA for image download
It's possible to specify instance_info['image_source'] to be a value
like 'glance://uuid'. When using AgentDeploy, image_source is passed
to IPA in image_info dictionary in 'id' field without modification,
and IPA tries to download image to /tmp/{id} location. Deployment
fails as /tmp/glance://uuid is invalid location.

In this change image_source is split to ensure that uuid will be used
to name the file.

Closes-bug: #1400156
Change-Id: Id3658d2a32c87867d828bc3b3fecc94e514a4526