Actively check iSCSI connection after login
This replaces the sleep command at the end of the login_iscsi fuction
with an active check and a forced lun update. These two things help ensure
the connection sees all luns on the iSCSI device before attempting to write
to it. It also adds a configurable option to allow users to be able to
customize the number of attempts for their environment.

Also add some very basic logging to aid in deploy debuging

Change-Id: I68c9f9f86f5f113bb111c0f4fd83216ae0659d36
Closes-bug: #1415117
Partial-Bug: #1286374