Add 'irmc_' prefix to optional properties
The optional driver_info properties are missing the 'irmc_'
prefix from their names. This patch adds them. The
descriptions were modified a bit so that 'Optional' is at the
end (easier for the reader to see/distinguish between Required
and Optional).

This is a follow-on patch to https://review.openstack.org/144901.
Implements: blueprint irmc-power-driver

Change-Id: I640ae0b4ec022b9cda6dd3f5194d4f2da4dc5935