Correctly rebuild the PXE file during takeover of ACTIVE nodes
During a conductor take-over, eg. from another failed conductor, it
rebuilds the PXE configuration of nodes that are mapped to it. This
calls node.driver.deploy.prepare(). However, in the PXE driver,
prepare() was not properly handling the case where it was called on a
node with an ACTIVE deploy state.

In such a situation, prepare() needs to switch from the "deploy" config
to the "boot" config.  Furthermore, the PXE boot file must specify the
node's root disk UUID.  This was previously not available outside of the
_continue_deploy() method, and so is now saved on
node.driver_internal_info when the initial deploy is completed.

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>
Co-Authored-By: Chris Krelle <nobodycam@gmail.com>
Change-Id: I6adc65b756e22c5c78d30231187fe2e8ccf2594d
Closes-bug: #1421370